Add EphemeralKDC test helper, refactor spike to use it
Extract the KDC lifecycle management from the Phase 1 spike into a
reusable EphemeralKDC struct in ephemeral_kdc_darwin_test.go:

- NewEphemeralKDC(t) handles the full lifecycle: locate MIT krb5 binaries,
  create temp directory, write kdc.conf/krb5.conf, initialize KDC database,
  create principals, start krb5kdc, and populate credential cache
- SetEnv(t) overrides KRB5_CONFIG and KRB5CCNAME for Apple Heimdal
- Close() stops the KDC process (also registered via t.Cleanup)

Helper functions: findKrb5Prefix, findFreePort, waitForPort, mustWriteFile,
mustRunCmd.

The spike test is refactored to use the new helper, reducing it from ~250
lines to ~50 lines.

Resolves Phase 2 of #108.
